+++ js/src/utils/v2_migration.ts
@@ -45,58 +45,21 @@ export function v2RunToRun(run: V2Run): Run {
   };
 }
 
-async function loadTracesV2(
+export async function loadTracesV2(
   client: Client,
   project: TracerSession,
-  options: { loadNested: boolean },
+  options: { isRoot?: boolean },
 ): Promise<Run[]> {
   const runs: Run[] = [];
   const pager = client.runs.queryV2({
     project_ids: [project.id],
     min_start_time: new Date(project.start_time).toISOString(),
     max_start_time: new Date(project.end_time ?? Date.now()).toISOString(),
-    is_root: options.loadNested ? undefined : true,
+    is_root: options.isRoot,
     selects: V2_RUN_SELECTS,
   });
   for await (const run of pager) {
     runs.push(v2RunToRun(run));
   }
   return runs;
-}
-
-export async function loadTraces(
-  client: Client,
-  project: TracerSession,
-  options: { loadNested: boolean },
-): Promise<Run[]> {
-  const runs = (await client._supportsSDBQuery())
-    ? await loadTracesV2(client, project, options)
-    : await client.listRuns({
-        projectId: project.id,
-        executionOrder: options.loadNested ? undefined : 1,
-      });
-
-  const treeMap: Record<string, Run[]> = {};
-  const runIdMap: Record<string, Run> = {};
-  const results: Run[] = [];
-
-  for await (const run of runs) {
-    if (run.parent_run_id != null) {
-      treeMap[run.parent_run_id] ??= [];
-      treeMap[run.parent_run_id].push(run);
-    } else {
-      results.push(run);
-    }
-    runIdMap[run.id] = run;
-  }
-
-  for (const [parentRunId, childRuns] of Object.entries(treeMap)) {
-    const parentRun = runIdMap[parentRunId];
-    parentRun.child_runs = childRuns.sort((a, b) => {
-      if (a.dotted_order == null || b.dotted_order == null) return 0;
-      return a.dotted_order.localeCompare(b.dotted_order);
-    });
-  }
-
-  return results;
 }
